requirements.txt
httpx>=0.24.0
uritemplate>=4.1.1
python-dotenv>=1.0.0

